http-otel/requirements.txt
flask
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-redis
